MAXQDA.

MAXQDA.

MAXQDA.

Details

Class to manage and get annotations from a MAXQDA file that stores annotations to html documents that have previously been generated by polmineR. These html documents are required to have character offset positions in tags.

Public fields

dbFilename

the MAXQDA file (*.mx12) from which annotations are to be extracted (full path)

codings

table with codings/annotations derived from the MAXQDA file, a data.table

codewords

table with codewords derived from the MAXQDA file, a data.table

texts

table with texts derived from the MAXQDA file, a data.table

htmlDir

directory where the html files that have been imported into MAXQDA reside

Methods

Public methods

MAXQDA$new()

MAXQDA$getMaxqdaData()

MAXQDA$getOffsetDT()

MAXQDA$getOne()

MAXQDA$getAll()

MAXQDA$clone()

Method new()

Create a new instance of class MAXQDA; the param dbFilename indicates the MAXQDA database, the htmlDir the directory where html files as generated with polmineR reside

Usage

MAXQDA$new(dbFilename, htmlDir)

Arguments

dbFilename

filename of the MAXQDA database (file ending: ".mx12")

htmlDir

html directory

Method getMaxqdaData()

Fill the fields codings, codewords, and texts with data.table by extracting the respective tables from the MAXQDA file specified by field dbFilename

Usage

MAXQDA$getMaxqdaData()

Method getOffsetDT()

Extract data.table with character offset positions and corpus positions from the html file specified by param filename

Usage

MAXQDA$getOffsetDT(filename)

Arguments

filename

filename of a html file that has been imported into MAXQDA

Method getOne()

Extract annotations for one document; essentially a worker called by the method getAll

Usage

MAXQDA$getOne(textId, offsetDT, corpus = "REUTERS")

Arguments

textId

DONTKNOW

offsetDT

DONTKNOW

corpus

corpus from which the partition is derived

Method getAll()

Extract annotations for all documents that reside in the htmlDir that has been specified upon initialization

Usage

MAXQDA$getAll(verbose = TRUE)

Arguments

verbose

logical, whether to be talkative

Method clone()

The objects of this class are cloneable with this method.

Usage

MAXQDA$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

maxfile <- system.file(package = "maxqda", "extdata", "maxqda", "reuters.mx12")
htmlDir <- system.file(package = "maxqda", "extdata", "html", "reuters")
M <- MAXQDA$new(dbFilename = maxfile, htmlDir = htmlDir)
annotationsDT <- M$getAll()

#> ... processing: argentina

#> ... processing: canada

#> ... processing: kuwait

#> ... processing: qatar